testing different StateID sizes

+++ src/nfa.zig
@@ -1,7 +1,7 @@
 const std = @import("std");
 const ast = @import("regex_ast.zig");
 
-pub const StateId = u16; // indices for states in NFA and DFA (determines max number of states)
+pub const StateId = usize; // indices for states in NFA and DFA (determines max number of states)
 
 pub const Transition = union(enum) {
     ch: struct { ch: u8, to: StateId },
